requirements.txt
asgiref==3.8.1
click==8.1.8
Django==5.1.7
django-cors-headers==4.7.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
geographiclib==2.0
geopy==2.4.1
gunicorn==23.0.0
h11==0.14.0
inflection==0.5.1
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
uvicorn==0.34.0
whitenoise==6.9.0
